[PRODUCT][EDIT][ACCESS] Update editing access to products - Product should be editable by admin as well as manager - Edit access should be given to owner or account admin - Add access related filters to mongo service - Block editing in editor dialog using transparent non clickable layer - Provide lead import access to managers

+++ grails-app/controllers/navimateforbusiness/api/ManagerApiController.groovy
@@ -836,8 +836,12 @@ class ManagerApiController {
                 throw new ApiException("Product not found...", Constants.HttpCodes.BAD_REQUEST)
             }
 
-            // Remove product
-            productService.remove(user, product)
+            // remove only if product is editable by this user
+            if (user.role == Role.ADMIN || product.ownerId == user.id) {
+                // Remove lead
+                reps.addAll(productService.getAffectedReps(user, product))
+                productService.remove(user, product)
+            }
         }
 
         def resp = [success: true]
